+++ tests/test_tags.py
@@ -19,8 +19,8 @@ def test_tag_init():
 
 def test_tag_characters():
 
-    good_tag = Tag("lettér 123_", "letter.123@something.com,characters456@blah.com and space")
-    assert good_tag.key == "lettér 123_"
+    good_tag = Tag("lettér123_", "letter.123@something.com,characters456@blah.com and space")
+    assert good_tag.key == "lettér123_"
     assert good_tag.value == "letter.123@something.com,characters456@blah.com and space"
 
     with pytest.raises(OAATemplateException) as e: